% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{friendlyeval_to_rlang}
\alias{friendlyeval_to_rlang}
\title{Convert friendlyeval functions to rlang}
\usage{
friendlyeval_to_rlang()
}
\value{
nothing.
}
\description{
Works on a RStudio document selection if one exists, or the entire
active source editor if no selection exists.
}
