% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyeval.R
\name{treat_input_as_col}
\alias{treat_input_as_col}
\title{treat_input_as_col}
\usage{
treat_input_as_col(arg)
}
\arguments{
\item{arg}{the argument for which the literal input text is to be used
as a column name.}
}
\value{
Something that will resolve to a column named when prefixed with !!.
}
\description{
Take what was input and treat it as a column name argument to a dplyr function.
}
\details{
This is used inside a function to pass the literal text of what the caller
typed as an argument to a `dplyr` function. When using `dplyr` the text will
typically be destined for a column name argument. See examples for usage
scenarios.
}
\examples{
\dontrun{
select_this <- function(a_col){
 select(mtcars, !!treat_input as_col(a_col))
}
select_this(cyl)

mean_this <- function(a_col){
 mutate(mtcars, result = mean(!!treat_input_as_col(a_col)))
}
mean_this(cyl)

filter_same <- function(dat, x, y) {
 filter(dat, !!treat_input_as_col(x) == !!treat_input_as_col(y))
}
filter_same(mtcars, carb, gear)
}
}
