% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyeval.R
\name{treat_inputs_as_cols}
\alias{treat_inputs_as_cols}
\title{treat_inputs_as_cols}
\usage{
treat_inputs_as_cols(...)
}
\arguments{
\item{...}{a comma separated list of arguments to be treated as column names.}
}
\value{
something that will resolve to a list of column names when prefixed
  with `!!!`
}
\description{
Take the literal text input for a comma separated list of arguments and treat it as list of column names in a dplyr function.
}
\details{
The most common usage of this is to pass `...`, from your function directly through to dplyr functions as column names, as in the `select_these` example.

This function must be prefixed with `!!!` to treat the output as a list.
}
\examples{
\dontrun{
select_these <- function(dat, ...){
 select(dat, !!!treat_inputs_as_cols(...))
}
select_these(mtcars, cyl, wt)
}
}
