% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyeval.R
\name{treat_inputs_as_exprs}
\alias{treat_inputs_as_exprs}
\title{treat_inputs_as_exprs}
\usage{
treat_inputs_as_exprs(...)
}
\arguments{
\item{...}{a comma separated list of arguments to be treated as expressions.}
}
\value{
something that will resolve to a list of expressions when prefixed with `!!!`
}
\description{
Take the literal text input for a comma separated list of arguments and treat it as a list of expressions in a dplyr function.
}
\details{
Common usage of this is to pass `...` from your function directly through to
dplyr functions as expressions. This could be a list of filtering
expressions for filter: `filter(dat, col1 == "example1", col2 ==
"example2")`, or a list of functions of columns to `mutate` or `summarise`:
`summarise(dat, mean(col1), var(col1))`

This function must be prefixed with `!!!` to treat the output as a list.
}
