% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyeval.R
\name{treat_string_as_col}
\alias{treat_string_as_col}
\title{treat_string_as_col}
\usage{
treat_string_as_col(arg)
}
\arguments{
\item{arg}{the argument that holds a value to be used as a column name.}
}
\value{
something that will resolve to a column name when prefixed with `!!`.
}
\description{
Take the a string value and use it as a column name in a
dplyr function.
}
\details{
This is used to take the string value of a variable and use it
in place of a literal column name when calling a dplyr function. This
ability is useful when the name of the column to operate on is determined at
run-time from data.
}
\examples{
\dontrun{
## drop this run-time determined column.
b <- "cyl"
 select(mtcars, -!!treat_string_as_col(b))

## function double a column
double_col <- function(dat, arg) {
  dplyr::mutate(dat, result = !!rlang::sym(arg) * 2)
}
double_col(mtcars, arg = 'cyl')

}
}
