% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyeval.R
\name{treat_string_as_expr}
\alias{treat_string_as_expr}
\title{treat_string_as_expr(x)}
\usage{
treat_string_as_expr(x)
}
\arguments{
\item{x}{a string to be treated as an expression.}
}
\value{
something that will resolve to an expression when prefixed with `!!`
}
\description{
Treat the string value of a variable as an expression in a dplyr function.
}
\details{
This will parse a string and treat it as an expression to be evaluated
in the context of a dplyr function call. This may be convenient when
building expressions to evaluate at run-time.
}
\examples{
\dontrun{

## processing operations from other notation 
calc_result <- function(dat, operation){
  operation <- gsub('x', '*', operation)
  mutate(dat, result = !!treat_string_as_expr(operation))
}

calc_result(mtcars, "mpg x hp")
}

}
