% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyeval.R
\name{typed_as_name}
\alias{typed_as_name}
\title{typed_as_name}
\usage{
typed_as_name(arg)
}
\arguments{
\item{a_name}{the argument for which the user typed text that is to be used
as a column name.}
}
\value{
Something that will resolve to a column named when prefixed with !!.
}
\description{
Take what was typed and use it as a column name argument in a dplyr
function.
}
\details{
This is used inside a function to pass the literal text of what the user
typed as an argument to another function. When using dplyr the text will
typically be destined for a column name argument. See examples for usage
scenarios.
}
\examples{
select_this <- function(a_col){
 select(mtcars, !!typed_as_name(a_col))
}
select_this(cyl)

mutate_this <- function(a_col){
 mutate(mtcars, result = mean(!!typed_as_name(a_col)))
}
mutate_this(cyl)

}
