% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyeval.R
\name{typed_as_name_lhs}
\alias{typed_as_name_lhs}
\title{typed_as_name_lhs}
\usage{
typed_as_name_lhs(arg)
}
\arguments{
\item{a_name}{the argument to convert to a column name}
}
\value{
Something that will resolve to a column name when prefixed with `!!`
}
\description{
Take what was typed and use it as a column name argument in dplyr on the
left hand side of an equality.
}
\details{
This is used inside a function to pass the literal text of what the user
typed as a function argument to another function. This function applies in
the special case that the text is intended to be used on the left hand side
of an an equality. For example to replace `my_col` in `mutate(my_col = pi)`.
In this case the call must be rewritten as
`mutate(!!typed_as_name_lhs(my_col) := pi)`. Note the usage of `:=`. This is
an additional requirement when using !! on the left hand side of an
equality.
}
\examples{

my_mutate1 <- function(dat, col_name){

mutate(dat,
      !!typed_as_name_lhs(colname) := 1
      )
}

mtcars \%>\%
  my_mutate1(cyl)
}
