% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlung.R
\name{typed_list_as_name_list}
\alias{typed_list_as_name_list}
\title{typed_list_as_name_list}
\usage{
typed_list_as_name_list(..., .named = FALSE, .ignore_empty = c("trailing",
  "none", "all"), .unquote_names = TRUE)
}
\arguments{
\item{a_list}{list of column names.}
}
\value{
something that will resolve to a list of column names when prefixed
  with `!!!`
}
\description{
Take what was typed for a comma separated list of parameters and pass it to
a another function as a comma separated list of column names
}
\details{
This is used to pass the literal text the user typed for a list of function
parameters and pass it to a dplyr function expecting a list of column names.
The most common usage of this is to pass `...`, as in the example provided.

This function must be prefixed with `!!!` to declare the output is a list.
}
\examples{
select_these <- function(dat, ...){
 select(dat, !!!typed_list_as_name_list(...))
}
select_these(mtcars, cyl, wt)
}
