% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlung.R
\name{value_as_name}
\alias{value_as_name}
\title{value_as_name}
\usage{
value_as_name(x)
}
\arguments{
\item{a_value}{}
}
\value{
something that will resolve to a column name when prefixed with `!!`.
}
\description{
Take the value of a character argument and use it as a column name in a
dplyr expression.
}
\details{
This is used to take the character value of a function argument and use it
in place of a literal column name when calling a dplyr function. This
ability is useful when the name of the column to operate on is determined at
run-time.
}
\examples{
b <- "cyl"
mtcars \%>\%
 select(-!!value_as_name(b))
}
