% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendlyeval.R
\name{value_list_as_name_list}
\alias{value_list_as_name_list}
\title{value_list_as_name_list}
\usage{
value_list_as_name_list(arg)
}
\arguments{
\item{arg}{the argument that holds a list of values to be used as column names.}
}
\value{
something that will resolve to a comma separated list of column
  names when prefixed with `!!!`.
}
\description{
Take a list of characters or character vectors and use the values as column
names.
}
\details{
This is used to take the character values of a list or vector supplied as a
function argument and use them as a comma separated list of column names in
a dplyr function. The most common usage would be to used on the values of a
single argument, as in the `select_these2` example, however it can also be
used with ... to transform all ... values to column names - see
`select_these3` example.
}
\examples{
\dontrun{
select_these2 <- function(dat, cols){
  select(dat, !!!value_list_as_name_list(cols))
}
select_these2(mtcars, cols = c("cyl", "wt"))

select_these3 <- function(dat, ...){
 dots <- list(...)
 select(dat, !!!value_list_as_name_list(dots))
}
select_these3(mtcars, "cyl", "wt")

select_not_these <- function(dat, cols){
  select(dat, -c(!!!value_list_as_name_list(cols)))
}
select_not_these(mtcars, cols = c("cyl", "wt"))
}
}
